Add files via upload
The Aim  of this project was Simple to show how   chatbot is useful.The chatbot should be easy and unfriendly to use,and respond in less time to showcase the information that user wants . The bots should make the users interaction as easy and fast as possible to ensure that the users precious time is not wasted and that they can retrieve information that they desire without any difficulty or misunderstanding from the bot. Chat-Bot Instruction should be easy to run user experience should be satisfied with the use of chat-bot and without any hesitation or wasting time user should be knowledgeable with the usage of chatbot. 


My contribution in completion of the project are as follows:
• Planing the strategy to create chat-bot
 • Build of the chat bot.
• Write a code in python to display information  through API.
• Displaying the retrieved information in the console
• Testing of the API displaying correct information to user 
• Experimented on different commands.